Position propeties after methods.
Properties are added (as class attributes) after functions to avoid conflicts
between propeties names and class names. For example in this stub definition:

class Window:
    Rect: Rect

    def GetClientRect(self) -> Rect: ...

mypy complains that the return type of GetClientRect is wx.Window.Rect, which
is not a type. Moving the properties after the methods solves this issue.

Fixes #4.